move all ikea specific converters to lib/ikea.ts
